sldj

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
   *
   * Return: n
   */
-int print_last_digit(int n)
+int print_last_digit(long long int n)
 {
 	int last_digit;
